refactor: tests폴더 src 폴더 밖으로 이동
