+++ Frontend/tests/TravelPlan.test.js
@@ -1,11 +1,11 @@
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
-import TravelPlan from "../pages/TravelPlan";
+import TravelPlan from "../src/pages/TravelPlan";
 import "@testing-library/jest-dom";
 import axios from "axios";
 import { BrowserRouter } from "react-router-dom";
 
 jest.mock("axios");
-jest.mock("../components/GoogleMapView", () => {
+jest.mock("./src/components/GoogleMapView", () => {
   return jest.fn(() => <div>Mock Google Map View</div>); // Render a simple div
 });
